requirements.txt
fastapi
uvicorn
jinja2
numpy
soundfile
SpeechRecognition
google-generativeai
gTTS
